optimized _zoomSurfaceY() for 2x and 4x cases
This covers the default resolution of 640x400 and the obvious next
choice of 1280x800. The new optimized functions use 64-bit integers
rather than reading pixels a byte at a time and writing them back
one byte a time. The result is an fps boost of up to x1.5. The optimized
functions also fall below TileEngine routines in CPU use according to
the callgrind profiler.